fix: separate discreate transformation from bbs files
